admin_panel/requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
python-dotenv